Convert all bigint database columns to integer. Refactor the Java code accordingly. The columns were originally bigint due to misinterpreting HealthMap's MySQL data type "int(11)" (this does NOT mean an integer with up to 11 significant figures, the 11 refers to how the integer is displayed).

+++ src/Common/src/uk/ac/ox/zoo/seeg/abraid/mp/common/dao/AlertDaoImpl.java
@@ -22,7 +22,7 @@ public class AlertDaoImpl extends AbstractDao<Alert, Integer> implements AlertDa
      * @return The alert, or null if not found.
      */
     @Override
-    public Alert getByHealthMapAlertId(Long healthMapAlertId) {
+    public Alert getByHealthMapAlertId(Integer healthMapAlertId) {
         return uniqueResultNamedQuery("getAlertByHealthMapAlertId", "healthMapAlertId", healthMapAlertId);
     }
 }
